+++ wNim/private/wFrame.nim
@@ -385,82 +385,6 @@ proc wFrame_OnMenuHighlight(event: wEvent) =
     self.mStatusBar.setStatusText(if selectedItem != nil: selectedItem.mHelp else: "")
     processed = true
 
-proc wFrame_OnMenuCommand(event: wEvent) =
-  let self = event.mWindow
-  var processed = false
-  defer: event.skip(if processed: false else: true)
-
-  let
-    pos = int event.mWparam
-    hmenu = cast[HMENU](event.mLparam)
-
-  var menuInfo = MENUINFO(cbSize: sizeof(MENUINFO), fMask: MIM_MENUDATA)
-  GetMenuInfo(hmenu, menuInfo)
-  if menuInfo.dwMenuData != 0:
-    let
-      menu = cast[wMenu](menuInfo.dwMenuData)
-      item = menu.mItemList[pos]
-
-    if item.mKind == wMenuItemCheck:
-      menu.toggle(pos)
-
-    elif item.mKind == wMenuItemRadio:
-      menu.check(pos)
-
-    # convet to wEvent_Menu message.
-    processed = self.processMessage(wEvent_Menu, cast[WPARAM](item.mId), 0,
-      event.mResult)
-
-when defined(useWinXP):
-  # under Windows XP, menu icon must draw by outself
-  proc wFrame_OnMeasureItem(event: wEvent) =
-    var processed = false
-    defer: event.skip(if processed: false else: true)
-
-    var pStruct = cast[LPMEASUREITEMSTRUCT](event.mLparam)
-    if pStruct.CtlType == ODT_MENU and pStruct.itemData != 0:
-      # here pStruct.itemData maybe a wMenu or a wMenuItem
-      let
-        menu = cast[wMenu](pStruct.itemData)
-        bmp = (if IsMenu(menu.mHmenu): menu.mBitmap else: cast[wMenuItem](pStruct.itemData).mBitmap)
-        iconHeight = GetSystemMetrics(SM_CYMENUSIZE)
-        iconWidth = GetSystemMetrics(SM_CXMENUSIZE)
-
-      if bmp != nil:
-        pStruct.itemHeight = max(bmp.mHeight + 2, iconHeight)
-        pStruct.itemWidth = max(bmp.mWidth + 4, iconWidth)
-        event.mResult = TRUE
-        processed = true
-
-  proc wFrame_OndrawItem(event: wEvent) =
-    var processed = false
-    defer: event.skip(if processed: false else: true)
-
-    var pStruct = cast[LPDRAWITEMSTRUCT](event.mLparam)
-    if pStruct.CtlType == ODT_MENU and pStruct.itemData != 0:
-      let
-        menu = cast[wMenu](pStruct.itemData)
-        bmp = (if IsMenu(menu.mHmenu): menu.mBitmap else: cast[wMenuItem](pStruct.itemData).mBitmap)
-
-      if bmp != nil:
-        let
-          width = bmp.mWidth
-          height = bmp.mHeight
-          memdc = CreateCompatibleDC(0)
-          prev = SelectObject(memdc, bmp.mHandle)
-          x = (pStruct.rcItem.right - pStruct.rcItem.left - width) div 2
-          y = (pStruct.rcItem.bottom - pStruct.rcItem.top - height) div 2
-
-        var bf = BLENDFUNCTION(BlendOp: AC_SRC_OVER, SourceConstantAlpha: 255,
-          AlphaFormat: AC_SRC_ALPHA)
-
-        AlphaBlend(pStruct.hDC, x, y, width, height, memdc, 0, 0, width, height, bf)
-
-        SelectObject(memdc, prev)
-        DeleteDC(memdc)
-        event.mResult = TRUE
-        processed = true
-
 proc final*(self: wFrame) =
   ## Default finalizer for wFrame.
   discard
@@ -478,14 +402,8 @@ proc init*(self: wFrame, owner: wWindow = nil, title = "", pos = wDefaultPoint,
     bgColor=GetSysColor(COLOR_APPWORKSPACE))
 
   self.systemConnect(wEvent_Size, wFrame_DoSize)
-
   self.hardConnect(wEvent_SetFocus, wFrame_OnSetFocus)
   self.hardConnect(wEvent_MenuHighlight, wFrame_OnMenuHighlight)
-  self.hardConnect(WM_MENUCOMMAND, wFrame_OnMenuCommand)
-
-  when defined(useWinXP):
-    self.hardConnect(WM_MEASUREITEM, wFrame_OnMeasureItem)
-    self.hardConnect(WM_DRAWITEM, wFrame_OndrawItem)
 
 proc Frame*(owner: wWindow = nil, title = "", pos = wDefaultPoint,
     size = wDefaultSize, style: wStyle = wDefaultFrameStyle,